# test(perf): new CLI for running and comparing perf tests