--- a/__mocks__/colorette.js
+++ b/__mocks__/colorette.js
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2019 American Express Travel Related Services Company, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+function buildTag(name) {
+  return (txt = '') => `<${name}>${txt}</${name}>`;
+}
+
+const tags = Object.keys(jest.requireActual('colorette')).reduce((acc, tag) => ({
+  ...acc,
+  [tag]: buildTag(tag),
+}), {});
+
+module.exports = {
+  ...tags,
+  createColors: () => tags,
+};