## Description

Adds new CLI for running and comparing perf tests.

```sh
$ npm run perf -- --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js --help

npm run perf -- <command>

Commands:
  npm run perf -- clean                Clean up performance test results
  npm run perf -- compare              Compare two performance test results
  npm run perf -- explain [metric]     Print descriptions of metrics
  npm run perf -- monitor              Monitor One App performance
  npm run perf -- summarize <results>  Summarize performance test results
                                                              [aliases: summary]
  npm run perf -- test                 Run a performance test

Options:
  -h, --help  Show help                                                [boolean]
```

```sh
$ npm run perf -- clean --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js clean --help

npm run perf -- clean

Clean up performance test results

Options:
  -n, --dry-run  List files that would be deleted without deleting them[boolean]
  -f, --force    Force deletion of files                               [boolean]
  -q, --quiet    Suppress output                                       [boolean]
  -h, --help     Show help                                             [boolean]
```

```sh
$ npm run perf -- compare --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js compare --help

npm run perf -- compare

Compare two performance test results

Options:
  -c, --control        Control test ID or path to control test results
                                                             [string] [required]
  -v, --variant        Variant test ID or path to variant test results
                                                             [string] [required]
      --markdown       Print the results table in markdown
                                                      [boolean] [default: false]
  -d, --description    Include a description of each metric in the results table
                                                      [boolean] [default: false]
  -a, --alpha          The significance level of the test
                                                        [number] [default: 0.05]
  -p, --processors     processors to run on each column of data
  [array] [choices: "sum", "mean", "median", "mode", "min", "max", "p95", "p90"]
                                                       [default: ["mean","max"]]
  -r, --raw            Output metrics unformatted (unitless)
                                                      [boolean] [default: false]
  -m, --metrics        metrics to compare
               [array] [default: Run npm run perf -- explain --defaults to list]
  -C, --control-label  Label for control test in report
                                                   [string] [default: "Control"]
  -V, --variant-label  Label for variant test in report
                                                   [string] [default: "Variant"]
      --show-hidden    Show hidden options                             [boolean]
  -h, --help           Show help                                       [boolean]
```

```sh
$ npm run perf -- explain --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js explain --help

npm run perf -- explain [metric]

Print descriptions of metrics

Positionals:
  metric  name of the metric to explain                                 [string]

Options:
      --label     look up metric by label                               [string]
      --defaults  list the default metrics                             [boolean]
  -h, --help      Show help                                            [boolean]
```

```sh
$ npm run perf -- monitor --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js monitor --help

npm run perf -- monitor

Monitor One App performance

Options:
  -t, --target      Host of the target metrics service
                                 [string] [required] [default: "localhost:3005"]
  -b, --background  Run the monitoring in the background               [boolean]
  -l, --tail        Tail backgrounded monitoring                       [boolean]
  -k, --kill        Kill the backgrounded monitoring                   [boolean]
  -h, --help        Show help                                          [boolean]
```

```sh
$ npm run perf -- summarize --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js summarize --help

npm run perf -- summarize <results>

Summarize performance test results

Positionals:
  results  Test ID or path to test results                   [string] [required]

Options:
      --markdown     Print the results table in markdown
                                                      [boolean] [default: false]
  -d, --description  Include a description of each metric in the results table
                                                      [boolean] [default: false]
  -p, --processors   processors to run on each column of data
  [array] [choices: "sum", "mean", "median", "mode", "min", "max", "p95", "p90"]
                                           [default: ["min","mean","max","p95"]]
  -r, --raw          Output metrics unformatted (unitless)
                                                      [boolean] [default: false]
  -m, --metrics      metrics to compare
               [array] [default: Run npm run perf -- explain --defaults to list]
  -h, --help         Show help                                         [boolean]
```

```sh
$ npm run perf -- test --help

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js test --help

npm run perf -- test

Run a performance test

Options:
      --type        Type of performance test to run
       [string] [required] [choices: "load", "smoke", "soak", "spike", "stress"]
      --target      URL of the target service                           [string]
      --identifier  Additional value included in the test ID            [string]
      --k6args      Additional arguments to pass to k6 run              [string]
  -h, --help        Show help                                          [boolean]
```

```sh
$ npm run perf -- summarize smoke-6.12.1-bdbc8e2-1713265335227

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js summarize smoke-6.12.1-bdbc8e2-1713265335227

┌────────────────────────────────────┐
│ smoke-6.12.1-bdbc8e2-1713265335227 │
├────────────────────────────────────┤
│ Test type: Smoke                   │
│ Completed: Today at 4:02 AM        │
│ Duration 11 seconds                │
│ HTTP Requests: 11                  │
│ Data Received: 119.505kB           │
│ Failed Requests: 0                 │
└────────────────────────────────────┘

┌───────────────────────┬────────────┬────────────┬────────────┬─────────────────┐
│ Metric                │ Min        │ Average    │ Max        │ 95th Percentile │
├───────────────────────┼────────────┼────────────┼────────────┼─────────────────┤
│ Request Duration      │ 11.179ms   │ 13.843ms   │ 17.024ms   │ 16.161ms        │
│ Request Blocked       │ 2.928µs    │ 294.596µs  │ 3.199ms    │ 5.264µs         │
│ TTFB                  │ 11.031ms   │ 13.648ms   │ 16.768ms   │ 16.05ms         │
│ Request Receiving     │ 92.372µs   │ 172.937µs  │ 271.341µs  │ 235.034µs       │
│ Heap Total            │ 87.441MB   │ 87.441MB   │ 87.441MB   │ 87.441MB        │
│ Heap Used             │ 54.717MB   │ 55.506MB   │ 56.61MB    │ 56.61MB         │
│ Used Space            │ 5.549MB    │ 5.557MB    │ 5.563MB    │ 5.563MB         │
│ Event loop Lag        │ 5.115ms    │ 5.423ms    │ 5.856ms    │ 5.856ms         │
│ Process CPU           │ 6.061%     │ 6.076%     │ 6.096%     │ 6.096%          │
│ System CPU            │ 0.78%      │ 0.805%     │ 0.84%      │ 0.84%           │
│ User CPU              │ 5.255%     │ 5.27%      │ 5.281%     │ 5.281%          │
│ GC Rate               │ 14.421µs/s │ 21.441µs/s │ 31.269µs/s │ 31.269µs/s      │
│ GC Duration           │ 55ms       │ 63.437ms   │ 75.25ms    │ 75.25ms         │
│ Circuit Breaker Open  │ 0/s        │ 0/s        │ 0/s        │ 0/s             │
│ Circuit Breaker Close │ 0/s        │ 0/s        │ 0/s        │ 0/s             │
└───────────────────────┴────────────┴────────────┴────────────┴─────────────────┘
```

```sh
$ npm run perf -- summarize smoke-6.12.1-bdbc8e2-1713265335227 --markdown

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js summarize smoke-6.12.1-bdbc8e2-1713265335227 --markdown
```

| smoke-6.12.1-bdbc8e2-1713265335227 |
| ---------------------------------- |
| Test type: **Smoke**               |
| Completed: **Today at 4:02 AM**    |
| Duration **11 seconds**            |
| HTTP Requests: **11**              |
| Data Received: **119.505kB**       |
| Failed Requests: **0**             |

| Metric                | Min        | Average    | Max        | 95th Percentile |
| --------------------- | ---------- | ---------- | ---------- | --------------- |
| Request Duration      | 11.179ms   | 13.843ms   | 17.024ms   | 16.161ms        |
| Request Blocked       | 2.928µs    | 294.596µs  | 3.199ms    | 5.264µs         |
| TTFB                  | 11.031ms   | 13.648ms   | 16.768ms   | 16.05ms         |
| Request Receiving     | 92.372µs   | 172.937µs  | 271.341µs  | 235.034µs       |
| Heap Total            | 87.441MB   | 87.441MB   | 87.441MB   | 87.441MB        |
| Heap Used             | 54.717MB   | 55.506MB   | 56.61MB    | 56.61MB         |
| Used Space            | 5.549MB    | 5.557MB    | 5.563MB    | 5.563MB         |
| Event loop Lag        | 5.115ms    | 5.423ms    | 5.856ms    | 5.856ms         |
| Process CPU           | 6.061%     | 6.076%     | 6.096%     | 6.096%          |
| System CPU            | 0.78%      | 0.805%     | 0.84%      | 0.84%           |
| User CPU              | 5.255%     | 5.27%      | 5.281%     | 5.281%          |
| GC Rate               | 14.421µs/s | 21.441µs/s | 31.269µs/s | 31.269µs/s      |
| GC Duration           | 55ms       | 63.437ms   | 75.25ms    | 75.25ms         |
| Circuit Breaker Open  | 0/s        | 0/s        | 0/s        | 0/s             |
| Circuit Breaker Close | 0/s        | 0/s        | 0/s        | 0/s             |

```sh
$ npm run perf -- compare --control=spike-6.12.1-61d7a9f-1713277160607 --variant=spike-6.12.1-61d7a9f-1713277440024

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js compare --control=spike-6.12.1-61d7a9f-1713277160607 --variant=spike-6.12.1-61d7a9f-1713277440024

┌────────────────────────────────────┬────────────────────────────────────┐
│ Control (c)                        │ Variant (v)                        │
├────────────────────────────────────┼────────────────────────────────────┤
│ Test type: Spike                   │ Test type: Spike                   │
│ Completed: Today at 7:23 AM        │ Completed: Today at 7:27 AM        │
│ Duration 4 minutes                 │ Duration 4 minutes                 │
│ HTTP Requests: 77932               │ HTTP Requests: 80208               │
│ Data Received: 538.279MB           │ Data Received: 553.261MB           │
│ Failed Requests: 0                 │ Failed Requests: 0                 │
│ spike-6.12.1-61d7a9f-1713277160607 │ spike-6.12.1-61d7a9f-1713277440024 │
└────────────────────────────────────┴────────────────────────────────────┘

┌───────────────────────┬─────────────┬─────────────┬────────────────────────┬─────────────┬─────────────┬────────────────────────┬─────────┐
│ Metric                │ Average (c) │ Average (v) │ Average (diff)         │ Max (c)     │ Max (v)     │ Max (diff)             │ p-value │
├───────────────────────┼─────────────┼─────────────┼────────────────────────┼─────────────┼─────────────┼────────────────────────┼─────────┤
│ Request Duration      │ 886.914ms   │ 868.93ms    │ -17.984ms (-2.028%)    │ 6.16s       │ 6.3s        │ +143.476ms (+2.33%)    │ *0.001  │
│ Request Blocked       │ 53.242ms    │ 52.051ms    │ -1.191ms (-2.236%)     │ 6.16s       │ 6.13s       │ -30.005ms (-0.487%)    │ 0.594   │
│ TTFB                  │ 886.799ms   │ 868.827ms   │ -17.972ms (-2.027%)    │ 6.16s       │ 6.3s        │ +143.577ms (+2.331%)   │ *0.001  │
│ Request Receiving     │ 92.265µs    │ 89.336µs    │ -2.929µs (-3.174%)     │ 25.781ms    │ 22.525ms    │ -3.256ms (-12.629%)    │ *0.009  │
│ Heap Total            │ 328.199MB   │ 333.858MB   │ +5.659MB (+1.724%)     │ 641.933MB   │ 537.35MB    │ -104.583MB (-16.292%)  │ 0.618   │
│ Heap Used             │ 205.6MB     │ 193.306MB   │ -12.293MB (-5.979%)    │ 598.694MB   │ 507.381MB   │ -91.313MB (-15.252%)   │ 0.255   │
│ Used Space            │ 6.593MB     │ 6.628MB     │ +35.624kB (+0.54%)     │ 6.868MB     │ 6.951MB     │ +83.2kB (+1.211%)      │ 0.121   │
│ Event loop Lag        │ 345.751ms   │ 229.596ms   │ -116.155ms (-33.595%)  │ 3.35s       │ 3.1s        │ -254.535ms (-7.598%)   │ 0.092   │
│ Process CPU           │ 108.965%    │ 111.532%    │ +2.567% (+2.356%)      │ 131.959%    │ 136.935%    │ +4.977% (+3.771%)      │ 0.404   │
│ System CPU            │ 11.56%      │ 12.097%     │ +0.538% (+4.652%)      │ 16.094%     │ 18.124%     │ +2.03% (+12.612%)      │ 0.091   │
│ User CPU              │ 97.406%     │ 99.435%     │ +2.029% (+2.083%)      │ 117.095%    │ 118.811%    │ +1.716% (+1.466%)      │ 0.473   │
│ GC Rate               │ 54.259µs/s  │ 68.221µs/s  │ +13.963µs/s (+25.734%) │ 124.838µs/s │ 164.576µs/s │ +39.738µs/s (+31.832%) │ *0.002  │
│ GC Duration           │ 11.716ms    │ 17.039ms    │ +5.322ms (+45.428%)    │ 59.5ms      │ 56.125ms    │ -3.375ms (-5.672%)     │ *<0.001 │
│ Circuit Breaker Open  │ 0.059/s     │ 0.056/s     │ -0.002/s (-4.06%)      │ 0.08/s      │ 0.08/s      │ -0/s (-0.188%)         │ 0.332   │
│ Circuit Breaker Close │ 0.038/s     │ 0.044/s     │ +0.006/s (+15.465%)    │ 0.08/s      │ 0.08/s      │ -0/s (-0.006%)         │ *<0.001 │
└───────────────────────┴─────────────┴─────────────┴────────────────────────┴─────────────┴─────────────┴────────────────────────┴─────────┘
```

```sh
$ npm run perf -- compare --control=spike-6.12.1-61d7a9f-1713277160607 --variant=spike-6.12.1-61d7a9f-1713277440024 --markdown

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js compare --control=spike-6.12.1-61d7a9f-1713277160607 --variant=spike-6.12.1-61d7a9f-1713277440024 --markdown
```

| Control (c)                        | Variant (v)                        |
| ---------------------------------- | ---------------------------------- |
| Test type: **Spike**               | Test type: **Spike**               |
| Completed: **Today at 7:23 AM**    | Completed: **Today at 7:27 AM**    |
| Duration **4 minutes**             | Duration **4 minutes**             |
| HTTP Requests: **77932**           | HTTP Requests: **80208**           |
| Data Received: **538.279MB**       | Data Received: **553.261MB**       |
| Failed Requests: **0**             | Failed Requests: **0**             |
| spike-6.12.1-61d7a9f-1713277160607 | spike-6.12.1-61d7a9f-1713277440024 |

| Metric                | Average (c) | Average (v) | Average (diff)         | Max (c)     | Max (v)     | Max (diff)             | _p_-value   |
| --------------------- | ----------- | ----------- | ---------------------- | ----------- | ----------- | ---------------------- | ----------- |
| Request Duration      | 886.914ms   | 868.93ms    | -17.984ms (-2.028%)    | 6.16s       | 6.3s        | +143.476ms (+2.33%)    | *0.001      |
| Request Blocked       | 53.242ms    | 52.051ms    | -1.191ms (-2.236%)     | 6.16s       | 6.13s       | -30.005ms (-0.487%)    | 0.594       |
| TTFB                  | 886.799ms   | 868.827ms   | -17.972ms (-2.027%)    | 6.16s       | 6.3s        | +143.577ms (+2.331%)   | *0.001      |
| Request Receiving     | 92.265µs    | 89.336µs    | -2.929µs (-3.174%)     | 25.781ms    | 22.525ms    | -3.256ms (-12.629%)    | *0.009      |
| Heap Total            | 328.199MB   | 333.858MB   | +5.659MB (+1.724%)     | 641.933MB   | 537.35MB    | -104.583MB (-16.292%)  | 0.618       |
| Heap Used             | 205.6MB     | 193.306MB   | -12.293MB (-5.979%)    | 598.694MB   | 507.381MB   | -91.313MB (-15.252%)   | 0.255       |
| Used Space            | 6.593MB     | 6.628MB     | +35.624kB (+0.54%)     | 6.868MB     | 6.951MB     | +83.2kB (+1.211%)      | 0.121       |
| Event loop Lag        | 345.751ms   | 229.596ms   | -116.155ms (-33.595%)  | 3.35s       | 3.1s        | -254.535ms (-7.598%)   | 0.092       |
| Process CPU           | 108.965%    | 111.532%    | +2.567% (+2.356%)      | 131.959%    | 136.935%    | +4.977% (+3.771%)      | 0.404       |
| System CPU            | 11.56%      | 12.097%     | +0.538% (+4.652%)      | 16.094%     | 18.124%     | +2.03% (+12.612%)      | 0.091       |
| User CPU              | 97.406%     | 99.435%     | +2.029% (+2.083%)      | 117.095%    | 118.811%    | +1.716% (+1.466%)      | 0.473       |
| GC Rate               | 54.259µs/s  | 68.221µs/s  | +13.963µs/s (+25.734%) | 124.838µs/s | 164.576µs/s | +39.738µs/s (+31.832%) | *0.002      |
| GC Duration           | 11.716ms    | 17.039ms    | +5.322ms (+45.428%)    | 59.5ms      | 56.125ms    | -3.375ms (-5.672%)     | ***<0.001** |
| Circuit Breaker Open  | 0.059/s     | 0.056/s     | -0.002/s (-4.06%)      | 0.08/s      | 0.08/s      | -0/s (-0.188%)         | 0.332       |
| Circuit Breaker Close | 0.038/s     | 0.044/s     | +0.006/s (+15.465%)    | 0.08/s      | 0.08/s      | -0/s (-0.006%)         | ***<0.001** |

```sh
$ npm run perf -- explain http_req_sending

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js explain http_req_sending

┌──────────────────────────────────────────────────────────────┐
│                       http_req_sending                       │
├──────────────────────────────────────────────────────────────┤
│                    Label: Request Sending                    │
│                                                              │
│   Description: time spent sending data to the remote host    │
└──────────────────────────────────────────────────────────────┘
```

```sh
$ npm run perf -- explain --label=ttfb

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js explain --label=ttfb

┌──────────────────────────────────────────────────────────────┐
│                       http_req_waiting                       │
├──────────────────────────────────────────────────────────────┤
│                         Label: TTFB                          │
│                                                              │
│   Description: time spent waiting for response from remote   │
│        host (a.k.a. “time to first byte”, or “TTFB”)         │
└──────────────────────────────────────────────────────────────┘

```

```sh
$ npm run perf -- explain

> @americanexpress/one-app@6.12.1 perf
> node __performance__/bin/index.js explain

┌───────────────────────┬──────────────────────────────────────────────────┬────────────────────────────────────────────────────┐
│ Label                 │ Metric                                           │ Description                                        │
├───────────────────────┼──────────────────────────────────────────────────┼────────────────────────────────────────────────────┤
│ HTTP Requests         │ http_reqs                                        │ How many total HTTP requests k6 generated          │
…